// src/Components/UI/Typography/enums.ts
export enum TypographyFonts {
  agrandir = 'font-agrandir',
  dmSans = 'font-dm-sans',
  spaceMono = 'font-space-mono'
}

export enum TypographySizes {
  xxs = 'font-size-xxs', // 12px
  xs = 'font-size-xs', // 14px
  s = 'font-size-s', // 16px
  sm = 'font-size-sm', // 18px
  m = 'font-size-m', // 20px
  m2 = 'font-size-m-2', // 22px
  xm = 'font-size-xm', // 26px
  xm2 = 'font-size-xm-2', // 28px
  xl2 = 'font-size-xl-2', // 32px
  xxl = 'font-size-xxl' // 50px
}

export enum TypographyColors {
  white = 'color-white',
  black = 'color-black',
  purple = 'color-purple',
  orange = 'color-orange',
  green = 'color-green',
  gray = 'color-gray'
}

export enum TypographyWeights {
  thin = 'weight-300',
  regular = 'weight-400',
  medium = 'weight-500',
  semibold = 'weight-600',
  bold = 'weight-700',
  extrabold = 'weight-800'
}